Removing WebServerManager Fix tests
