+++ Projects/Dotmim.Sync.Web.Server/WebServerOrchestrator.cs
@@ -5,6 +5,7 @@ using Microsoft.AspNetCore.Http;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 using System;
+using System.Collections.Generic;
 using System.IO;
 using System.IO.Compression;
 using System.Linq;
@@ -61,11 +62,9 @@ namespace Dotmim.Sync.Web.Server
             var serAndsizeString = string.Empty;
             var cliConverterKey = string.Empty;
 
-            // Get the serialization and batch size format
             if (TryGetHeaderValue(httpContext.Request.Headers, "dotmim-sync-serialization-format", out var vs))
                 serAndsizeString = vs.ToLowerInvariant();
 
-            // Get the serialization and batch size format
             if (TryGetHeaderValue(httpContext.Request.Headers, "dotmim-sync-converter", out var cs))
                 cliConverterKey = cs.ToLowerInvariant();
 
@@ -334,6 +333,22 @@ namespace Dotmim.Sync.Web.Server
                 throw new HttpConverterNotConfiguredException(this.WebServerOptions.Converters.Select(sf => sf.Key));
             }
         }
+
+
+        /// <summary>
+        /// Get Scope Name sent by the client
+        /// </summary>
+        public static string GetScopeName(HttpContext httpContext) => TryGetHeaderValue(httpContext.Request.Headers, "dotmim-sync-scope-name", out var val) ? val : null;
+
+        /// <summary>
+        /// Get the current client session id
+        /// </summary>
+        public static string GetClientSessionId(HttpContext httpContext) => TryGetHeaderValue(httpContext.Request.Headers, "dotmim-sync-session-id", out var val) ? val : null;
+        
+        /// <summary>
+        /// Get the current Step
+        /// </summary>
+        public static HttpStep GetCurrentStep(HttpContext httpContext) => TryGetHeaderValue(httpContext.Request.Headers, "dotmim-sync-step", out var val) ? (HttpStep)Convert.ToInt32(val) : HttpStep.None;
 
         public static bool TryGetHeaderValue(IHeaderDictionary n, string key, out string header)
         {
@@ -992,6 +1007,134 @@ namespace Dotmim.Sync.Web.Server
             await httpResponse.Body.WriteAsync(compressedData, 0, compressedData.Length, default).ConfigureAwait(false);
 
         }
+
+        public static Task WriteHelloAsync(HttpContext context, WebServerOrchestrator orchestrator, CancellationToken cancellationToken = default)
+                    => WriteHelloAsync(context, new[] { orchestrator }, cancellationToken);
+
+        public static async Task WriteHelloAsync(HttpContext context, IEnumerable<WebServerOrchestrator> orchestrators, CancellationToken cancellationToken = default)
+        {
+            var httpResponse = context.Response;
+            var stringBuilder = new StringBuilder();
+
+
+            stringBuilder.AppendLine("<!doctype html>");
+            stringBuilder.AppendLine("<html>");
+            stringBuilder.AppendLine("<head>");
+            stringBuilder.AppendLine("<meta charset='utf-8'>");
+            stringBuilder.AppendLine("<meta name='viewport' content='width=device-width, initial-scale=1, shrink-to-fit=no'>");
+            stringBuilder.AppendLine("<script src='https://cdn.jsdelivr.net/gh/google/code-prettify@master/loader/run_prettify.js'></script>");
+            stringBuilder.AppendLine("<link rel='stylesheet' href='https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/css/bootstrap.min.css' integrity='sha384-Vkoo8x4CGsO3+Hhxv8T/Q5PaXtkKtu6ug5TOeNV6gBiFeWPGFN9MuhOf23Q9Ifjh' crossorigin='anonymous'>");
+            stringBuilder.AppendLine("</head>");
+            stringBuilder.AppendLine("<title>Web Server properties</title>");
+            stringBuilder.AppendLine("<body>");
+
+
+            stringBuilder.AppendLine("<div class='container'>");
+            stringBuilder.AppendLine("<h2>Web Server properties</h2>");
+
+            foreach (var webOrchestrator in orchestrators)
+            {
+
+                string dbName = null;
+                string version = null;
+                string exceptionMessage = null;
+                bool hasException = false;
+                try
+                {
+                    (dbName, version) = await webOrchestrator.GetHelloAsync();
+                }
+                catch (Exception ex)
+                {
+                    exceptionMessage = ex.Message;
+                    hasException = true;
+
+                }
+
+                stringBuilder.AppendLine("<ul class='list-group mb-2'>");
+                stringBuilder.AppendLine($"<li class='list-group-item active'>Trying to reach database</li>");
+                stringBuilder.AppendLine("</ul>");
+                if (hasException)
+                {
+                    stringBuilder.AppendLine("<ul class='list-group mb-2'>");
+                    stringBuilder.AppendLine($"<li class='list-group-item list-group-item-primary'>Exception occured</li>");
+                    stringBuilder.AppendLine($"<li class='list-group-item list-group-item-danger'>");
+                    stringBuilder.AppendLine($"{exceptionMessage}");
+                    stringBuilder.AppendLine("</li>");
+                    stringBuilder.AppendLine("</ul>");
+                }
+                else
+                {
+                    stringBuilder.AppendLine("<ul class='list-group mb-2'>");
+                    stringBuilder.AppendLine($"<li class='list-group-item list-group-item-primary'>Database</li>");
+                    stringBuilder.AppendLine($"<li class='list-group-item list-group-item-light'>");
+                    stringBuilder.AppendLine($"Check database {dbName}: Done.");
+                    stringBuilder.AppendLine("</li>");
+                    stringBuilder.AppendLine("</ul>");
+
+                    stringBuilder.AppendLine("<ul class='list-group mb-2'>");
+                    stringBuilder.AppendLine($"<li class='list-group-item list-group-item-primary'>Engine version</li>");
+                    stringBuilder.AppendLine($"<li class='list-group-item list-group-item-light'>");
+                    stringBuilder.AppendLine($"{version}");
+                    stringBuilder.AppendLine("</li>");
+                    stringBuilder.AppendLine("</ul>");
+                }
+
+                stringBuilder.AppendLine("<ul class='list-group mb-2'>");
+                stringBuilder.AppendLine($"<li class='list-group-item active'>ScopeName: {webOrchestrator.ScopeName}</li>");
+                stringBuilder.AppendLine("</ul>");
+
+                var s = JsonConvert.SerializeObject(webOrchestrator.Setup, Formatting.Indented);
+                stringBuilder.AppendLine("<ul class='list-group mb-2'>");
+                stringBuilder.AppendLine($"<li class='list-group-item list-group-item-primary'>Setup</li>");
+                stringBuilder.AppendLine($"<li class='list-group-item list-group-item-light'>");
+                stringBuilder.AppendLine("<pre class='prettyprint' style='border:0px;font-size:75%'>");
+                stringBuilder.AppendLine(s);
+                stringBuilder.AppendLine("</pre>");
+                stringBuilder.AppendLine("</li>");
+                stringBuilder.AppendLine("</ul>");
+
+                s = JsonConvert.SerializeObject(webOrchestrator.Provider, Formatting.Indented);
+                stringBuilder.AppendLine("<ul class='list-group mb-2'>");
+                stringBuilder.AppendLine($"<li class='list-group-item list-group-item-primary'>Provider</li>");
+                stringBuilder.AppendLine($"<li class='list-group-item list-group-item-light'>");
+                stringBuilder.AppendLine("<pre class='prettyprint' style='border:0px;font-size:75%'>");
+                stringBuilder.AppendLine(s);
+                stringBuilder.AppendLine("</pre>");
+                stringBuilder.AppendLine("</li>");
+                stringBuilder.AppendLine("</ul>");
+
+                s = JsonConvert.SerializeObject(webOrchestrator.Options, Formatting.Indented);
+                stringBuilder.AppendLine("<ul class='list-group mb-2'>");
+                stringBuilder.AppendLine($"<li class='list-group-item list-group-item-primary'>Options</li>");
+                stringBuilder.AppendLine($"<li class='list-group-item list-group-item-light'>");
+                stringBuilder.AppendLine("<pre class='prettyprint' style='border:0px;font-size:75%'>");
+                stringBuilder.AppendLine(s);
+                stringBuilder.AppendLine("</pre>");
+                stringBuilder.AppendLine("</li>");
+                stringBuilder.AppendLine("</ul>");
+
+                s = JsonConvert.SerializeObject(webOrchestrator.WebServerOptions, Formatting.Indented);
+                stringBuilder.AppendLine("<ul class='list-group mb-2'>");
+                stringBuilder.AppendLine($"<li class='list-group-item list-group-item-primary'>Web Server Options</li>");
+                stringBuilder.AppendLine($"<li class='list-group-item list-group-item-light'>");
+                stringBuilder.AppendLine("<pre class='prettyprint' style='border:0px;font-size:75%'>");
+                stringBuilder.AppendLine(s);
+                stringBuilder.AppendLine("</pre>");
+                stringBuilder.AppendLine("</li>");
+                stringBuilder.AppendLine("</ul>");
+
+
+
+            }
+            stringBuilder.AppendLine("</div>");
+            stringBuilder.AppendLine("</body>");
+            stringBuilder.AppendLine("</html>");
+
+            await httpResponse.WriteAsync(stringBuilder.ToString(), cancellationToken);
+
+
+        }
+
 
     }
 }
