+++ Samples/Batchsize/BatchsizeServer/Controllers/SyncController.cs
@@ -29,6 +29,6 @@ namespace BatchsizeServer.Controllers
         /// The configuration is shown only if Environmenent == Development
         /// </summary>
         [HttpGet]
-        public Task Get() => WebServerManager.WriteHelloAsync(this.HttpContext, WebServerOrchestrator);
+        public Task Get() => WebServerOrchestrator.WriteHelloAsync(this.HttpContext, WebServerOrchestrator);
     }
 }
