+++ Samples/FilterWebSync/FilterWebSyncServer/Controllers/SyncController.cs
@@ -31,7 +31,7 @@ namespace FilterWebSyncServer.Controllers
         /// The configuration is shown only if Environmenent == Development
         /// </summary>
         [HttpGet]
-        public Task Get() => WebServerManager.WriteHelloAsync(this.HttpContext, WebServerOrchestrator);
+        public Task Get() => WebServerOrchestrator.WriteHelloAsync(this.HttpContext, WebServerOrchestrator);
 
     }
 }
