+++ Samples/HelloWebSync/HelloWebSyncServer/Controllers/SyncController.cs
@@ -32,7 +32,7 @@ namespace HelloWebSyncServer.Controllers
         /// The configuration is shown only if Environmenent == Development
         /// </summary>
         [HttpGet]
-        public Task Get() => WebServerManager.WriteHelloAsync(this.HttpContext, WebServerOrchestrator);
+        public Task Get() => WebServerOrchestrator.WriteHelloAsync(this.HttpContext, WebServerOrchestrator);
 
     }
 }
