+++ Samples/XamAppSync/XamAppServer/Controllers/SyncController.cs
@@ -28,6 +28,6 @@ namespace XamAppServer.Controllers
         /// The configuration is shown only if Environmenent == Development
         /// </summary>
         [HttpGet]
-        public Task Get() => WebServerManager.WriteHelloAsync(this.HttpContext, WebServerOrchestrator);
+        public Task Get() => WebServerOrchestrator.WriteHelloAsync(this.HttpContext, WebServerOrchestrator);
     }
 }
